additional, cast, protect

+++ libft/ft_strequ.c
@@ -10,11 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
-
 int		ft_strequ(char const *s1, char const *s2)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (!s1 && !s2)
